third_party/WebKit/LayoutTests/external/wpt/web-animations/interfaces/Animation/constructor-expected.txt

This is a testharness.js-based test.
FAIL Animation can be constructed with null effect and null timeline assert_equals: Animation timeline should be null expected null but got object "[object DocumentTimeline]"
PASS Animation can be constructed with null effect and non-null timeline
PASS Animation can be constructed with null effect and no timeline parameter
FAIL Animation can be constructed with non-null effect and null timeline assert_equals: Animation timeline should be null expected null but got object "[object DocumentTimeline]"
PASS Animation can be constructed with non-null effect and non-null timeline
PASS Animation can be constructed with non-null effect and no timeline parameter
PASS Animation constructed by an effect with null target runs normally
PASS Animation constructed with a keyframe that target element is in iframe
Harness: the test ran to completion.
